requirements.txt
fastapi
django
djangorestframework 
psycopg2-binary
uvicorn
python-dotenv
pyyaml